+++ beatmap/parse.ts
@@ -1,3 +1,3 @@
-export { difficulty as difficultyV1, info as infoV1 } from './v1/parse.ts';
-export { difficulty as difficultyV2, info as infoV2 } from './v2/parse.ts';
-export { difficulty as difficultyV3 } from './v3/parse.ts';
+export { difficulty as v1Difficulty, info as v1Info } from './v1/parse.ts';
+export { difficulty as v2Difficulty, info as v2Info } from './v2/parse.ts';
+export { difficulty as v3Difficulty } from './v3/parse.ts';