+++ patch/dataCorrection/customDataObject.ts
@@ -1,7 +1,7 @@
-import { ICustomDataNote as ICustomDataNoteV2 } from '../../types/beatmap/v2/custom/note.ts';
-import { ICustomDataObstacle as ICustomDataObstacleV2 } from '../../types/beatmap/v2/custom/obstacle.ts';
-import { ICustomDataNote as ICustomDataNoteV3 } from '../../types/beatmap/v3/custom/note.ts';
-import { ICustomDataObstacle as ICustomDataObstacleV3 } from '../../types/beatmap/v3/custom/obstacle.ts';
+import { ICustomDataNote as IV2CustomDataNote } from '../../types/beatmap/v2/custom/note.ts';
+import { ICustomDataObstacle as IV2CustomDataObstacle } from '../../types/beatmap/v2/custom/obstacle.ts';
+import { ICustomDataNote as IV3CustomDataNote } from '../../types/beatmap/v3/custom/note.ts';
+import { ICustomDataObstacle as IV3CustomDataObstacle } from '../../types/beatmap/v3/custom/obstacle.ts';
 import { ICustomDataSlider } from '../../types/beatmap/v3/custom/slider.ts';
 import { Vector3 } from '../../types/vector.ts';
 import {
@@ -19,10 +19,10 @@ import {
 
 export function fixCustomDataObject(
    cd?:
-      & ICustomDataNoteV2
-      & ICustomDataObstacleV2
-      & ICustomDataNoteV3
-      & ICustomDataObstacleV3
+      & IV2CustomDataNote
+      & IV2CustomDataObstacle
+      & IV3CustomDataNote
+      & IV3CustomDataObstacle
       & ICustomDataSlider,
 ) {
    if (!cd) {