+++ save.ts
@@ -10,16 +10,16 @@ import globals from './globals.ts';
 import logger from './logger.ts';
 import { deepCheck } from './beatmap/shared/dataCheck.ts';
 import {
-   DifficultyCheck as DifficultyCheckV1,
-   InfoCheck as InfoCheckV1,
+   DifficultyCheck as V1DifficultyCheck,
+   InfoCheck as IV1nfoCheck,
 } from './beatmap/v1/dataCheck.ts';
 import {
-   DifficultyCheck as DifficultyCheckV2,
-   InfoCheck as InfoCheckV2,
+   DifficultyCheck as V2DifficultyCheck,
+   InfoCheck as IV2nfoCheck,
 } from './beatmap/v2/dataCheck.ts';
-import { DifficultyCheck as DifficultyCheckV3 } from './beatmap/v3/dataCheck.ts';
-import { IDifficulty as IDifficultyV2 } from './types/beatmap/v2/difficulty.ts';
-import { Info as InfoV2 } from './beatmap/v2/info.ts';
+import { DifficultyCheck as V3DifficultyCheck } from './beatmap/v3/dataCheck.ts';
+import { IDifficulty as IV2Difficulty } from './types/beatmap/v2/difficulty.ts';
+import { Info as IV2nfo } from './beatmap/v2/info.ts';
 import { IWrapInfo } from './types/beatmap/wrapper/info.ts';
 import { IWrapDifficulty } from './types/beatmap/wrapper/difficulty.ts';
 import { resolve } from './deps.ts';
@@ -80,17 +80,17 @@ function _info(data: IWrapInfo, options: ISaveOptionsInfo) {
       validate: options.validate ?? defaultOptions.info.validate,
       dataCheck: options.dataCheck ?? defaultOptions.difficulty.dataCheck,
    };
-   const ver = data instanceof InfoV2 ? 2 : 1;
+   const ver = data instanceof IV2nfo ? 2 : 1;
    const objectData = data.toJSON();
    if (opt.optimize.enabled) {
       optimize.info(objectData as IInfo, opt.optimize);
    }
    if (opt.dataCheck.enabled) {
       if (ver === 1) {
-         deepCheck(objectData, InfoCheckV1, 'difficulty', '1.0.0', opt.dataCheck.throwError);
+         deepCheck(objectData, IV1nfoCheck, 'difficulty', '1.0.0', opt.dataCheck.throwError);
       }
       if (ver === 2) {
-         deepCheck(objectData, InfoCheckV2, 'difficulty', '2.2.0', opt.dataCheck.throwError);
+         deepCheck(objectData, IV2nfoCheck, 'difficulty', '2.2.0', opt.dataCheck.throwError);
       }
    }
    const p = resolve(opt.directory, opt.filePath);
@@ -147,17 +147,17 @@ function _difficulty(data: IWrapDifficulty, options: ISaveOptionsDifficulty) {
    const ver = data.version;
    const objectData = data.toJSON();
    if (opt.optimize.enabled) {
-      optimize.difficulty(objectData as IDifficultyV2, opt.optimize);
+      optimize.difficulty(objectData as IV2Difficulty, opt.optimize);
    }
    if (opt.dataCheck.enabled) {
       if (ver.startsWith('1')) {
-         deepCheck(objectData, DifficultyCheckV1, 'difficulty', ver, opt.dataCheck.throwError);
+         deepCheck(objectData, V1DifficultyCheck, 'difficulty', ver, opt.dataCheck.throwError);
       }
       if (ver.startsWith('2')) {
-         deepCheck(objectData, DifficultyCheckV2, 'difficulty', ver, opt.dataCheck.throwError);
+         deepCheck(objectData, V2DifficultyCheck, 'difficulty', ver, opt.dataCheck.throwError);
       }
       if (ver.startsWith('3')) {
-         deepCheck(objectData, DifficultyCheckV3, 'difficulty', ver, opt.dataCheck.throwError);
+         deepCheck(objectData, V3DifficultyCheck, 'difficulty', ver, opt.dataCheck.throwError);
       }
    }
    const p = resolve(opt.directory, opt.filePath);
@@ -219,13 +219,13 @@ function _difficultyList(difficulties: ILoadInfoData[], options: ISaveOptionsDif
       }
       if (opt.dataCheck.enabled) {
          if (ver.startsWith('1')) {
-            deepCheck(objectData, DifficultyCheckV1, 'difficulty', ver, opt.dataCheck.throwError);
+            deepCheck(objectData, V1DifficultyCheck, 'difficulty', ver, opt.dataCheck.throwError);
          }
          if (ver.startsWith('2')) {
-            deepCheck(objectData, DifficultyCheckV2, 'difficulty', ver, opt.dataCheck.throwError);
+            deepCheck(objectData, V2DifficultyCheck, 'difficulty', ver, opt.dataCheck.throwError);
          }
          if (ver.startsWith('3')) {
-            deepCheck(objectData, DifficultyCheckV3, 'difficulty', ver, opt.dataCheck.throwError);
+            deepCheck(objectData, V3DifficultyCheck, 'difficulty', ver, opt.dataCheck.throwError);
          }
       }
       const p = resolve(opt.directory, dl.settings.filename);