+++ types/bsmap/infoDiff.ts
@@ -1,9 +1,9 @@
 import { CharacteristicName } from '../beatmap/shared/characteristic.ts';
 import { DifficultyName } from '../beatmap/shared/difficulty.ts';
 import { IWrapInfoDifficulty } from '../beatmap/wrapper/info.ts';
-import { Difficulty as DifficultyV1 } from '../../beatmap/v1/difficulty.ts';
-import { Difficulty as DifficultyV2 } from '../../beatmap/v2/difficulty.ts';
-import { Difficulty as DifficultyV3 } from '../../beatmap/v3/difficulty.ts';
+import { Difficulty as V1Difficulty } from '../../beatmap/v1/difficulty.ts';
+import { Difficulty as V2Difficulty } from '../../beatmap/v2/difficulty.ts';
+import { Difficulty as V3Difficulty } from '../../beatmap/v3/difficulty.ts';
 import { IWrapDifficulty } from '../beatmap/wrapper/difficulty.ts';
 
 interface ILoadInfoDataBase {
@@ -13,19 +13,19 @@ interface ILoadInfoDataBase {
    data: IWrapDifficulty;
 }
 
-interface ILoadInfoDifficultyV1 extends ILoadInfoDataBase {
+interface IV1LoadInfoDifficulty extends ILoadInfoDataBase {
    version: 1;
-   data: DifficultyV1;
+   data: V1Difficulty;
 }
 
-interface ILoadInfoDifficultyV2 extends ILoadInfoDataBase {
+interface IV2LoadInfoDifficulty extends ILoadInfoDataBase {
    version: 2;
-   data: DifficultyV2;
+   data: V2Difficulty;
 }
 
-interface ILoadInfoDifficultyV3 extends ILoadInfoDataBase {
+interface IV3LoadInfoDifficulty extends ILoadInfoDataBase {
    version: 3;
-   data: DifficultyV3;
+   data: V3Difficulty;
 }
 
-export type ILoadInfoData = ILoadInfoDifficultyV1 | ILoadInfoDifficultyV2 | ILoadInfoDifficultyV3;
+export type ILoadInfoData = IV1LoadInfoDifficulty | IV2LoadInfoDifficulty | IV3LoadInfoDifficulty;